Phase 3.1c-iv-f: Compose TV hub is the only browse host
Delete Leanback RootBrowseFragment/BaseHttpBrowseFragment/CardPresenter
and Leanback card wrappers; drop leanback-preference and tv_main. Keep
leanback for VideoOverlay HorizontalGridView. Extract ImageCardContent
for the Compose hub.
