+++ app/AndroidManifest.xml
@@ -72,7 +72,7 @@
             android:exported="false" />
         <activity
             android:name=".tv.activities.PreferenceActivity"
-            android:theme="@style/Theme.Dreamdroid.LeanbackPreferences"
+            android:theme="@style/Theme.DreamDroid"
             android:exported="false" />
 
         <activity
