+++ app/build.gradle
@@ -36,8 +36,7 @@ dependencies {
     implementation 'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'androidx.leanback:leanback:1.0.0'
-    implementation 'androidx.leanback:leanback-preference:1.0.0'
-    // Phase 3.1c-iv-b: Compose for TV (hub beachhead; Leanback remains default hub).
+    // Phase 3.1c-iv: Compose for TV hub. Leanback kept for VideoOverlay HorizontalGridView.
     implementation 'androidx.tv:tv-foundation:1.0.0-alpha11'
     implementation 'androidx.tv:tv-material:1.0.0'
     implementation 'androidx.annotation:annotation:1.5.0'
