--- app/res/layout-television/tv_main.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main_frame"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:ignore="MissingDefaultResource">
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/main_browse_fragment"
-        android:name="net.reichholf.dreamdroid.tv.fragment.RootBrowseFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
-</FrameLayout>
