+++ app/res/values/themes.xml
@@ -128,13 +128,6 @@
         <item name="android:windowTranslucentNavigation">true</item>
     </style>
 
-    <style name="Theme.Dreamdroid.LeanbackPreferences" parent="Theme.Leanback">
-        <item name="preferenceTheme">@style/PreferenceThemeOverlayLeanback</item>
-        <item name="android:windowIsTranslucent">true</item>
-        <item name="android:windowBackground">@android:color/transparent</item>
-        <item name="android:backgroundDimEnabled">true</item>
-        <item name="android:colorPrimary">@color/primary_material_dark</item>
-    </style>
 
     <style name="VideoPlayerButton" parent="Base.Widget.AppCompat.Button.Borderless" />
