+++ app/src/net/reichholf/dreamdroid/tv/activities/PreferenceActivity.kt
@@ -32,7 +32,7 @@ class PreferenceActivity : ComponentActivity() {
         title = getString(if (isProfileMode) R.string.profile else R.string.settings)
 
         if (isProfileMode) {
-            // Match old ProfileFragment: RESULT_OK so RootBrowse reloads on return.
+            // Match old ProfileFragment: RESULT_OK so the TV hub reloads on return.
             setResult(Activity.RESULT_OK)
             val profile = DreamDroid.getCurrentProfile()
             editingProfile = profile
