--- app/src/net/reichholf/dreamdroid/tv/fragment/RootBrowseFragment.java
@@ -1,440 +1,0 @@
-package net.reichholf.dreamdroid.tv.fragment;
-
-import android.content.Intent;
-import android.os.Bundle;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.leanback.widget.ArrayObjectAdapter;
-import androidx.leanback.widget.HeaderItem;
-import androidx.leanback.widget.ListRow;
-import androidx.leanback.widget.ObjectAdapter;
-import androidx.leanback.widget.Presenter;
-import androidx.leanback.widget.Row;
-import androidx.leanback.widget.RowPresenter;
-import androidx.core.content.ContextCompat;
-import android.widget.Toast;
-
-import net.reichholf.dreamdroid.DreamDroid;
-import net.reichholf.dreamdroid.Profile;
-import net.reichholf.dreamdroid.ProfileChangedListener;
-import net.reichholf.dreamdroid.R;
-import net.reichholf.dreamdroid.helpers.ExtendedHashMap;
-import net.reichholf.dreamdroid.helpers.NameValuePair;
-import net.reichholf.dreamdroid.enigma.EpgNowNextLoadKt;
-import net.reichholf.dreamdroid.enigma.MovieListLoadKt;
-import net.reichholf.dreamdroid.enigma.ServiceListLoadKt;
-import net.reichholf.dreamdroid.enigma.TvBrowsePrefetchKt;
-import net.reichholf.dreamdroid.enigma.Service;
-import net.reichholf.dreamdroid.enigma.ServiceNowNext;
-import net.reichholf.dreamdroid.intents.IntentFactory;
-import net.reichholf.dreamdroid.ui.services.MovieListMapperKt;
-import net.reichholf.dreamdroid.ui.services.ServiceListMapperKt;
-import net.reichholf.dreamdroid.tv.BrowseItem;
-import net.reichholf.dreamdroid.tv.activities.PreferenceActivity;
-import net.reichholf.dreamdroid.tv.fragment.abs.BaseHttpBrowseFragment;
-import net.reichholf.dreamdroid.tv.presenter.CardPresenter;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-
-import kotlinx.coroutines.Job;
-
-/**
- * Created by Stephan on 16.10.2016.
- */
-
-public class RootBrowseFragment extends BaseHttpBrowseFragment implements ProfileChangedListener {
-
-	public static final int LOADER_BOUQUETLIST_ID = 1;
-	public static final int LOADER_SERVICELIST_ID = 2;
-	public static final int LOADER_MOVIELIST_ID = 3;
-
-	public static final int REQUEST_CODE_PROFILE = 0x1337;
-
-	public static final int ROW_SETTINGS_ID = 65535;
-	public static final int ROW_MOVIE_OFFSET_ID = 0x1000;
-
-	@NonNull
-	public static String BOUQUETS_TV = "1:7:1:0:0:0:0:0:0:0:(type == 1) || (type == 17) || (type == 195) || (type == 25) FROM BOUQUET \\\"bouquets.tv\\\" ORDER BY bouquet";
-
-	boolean mRequireReload;
-
-	@NonNull
-	ArrayList<ArrayObjectAdapter> mListRowAdapters = new ArrayList<>();
-
-	@Nullable
-	ArrayList<Service> mBouquets = null;
-	@Nullable
-	ArrayList<Service> mBouquetQueue = null;
-
-	HashMap<String, ArrayList<net.reichholf.dreamdroid.enigma.Movie>> mLocations;
-	@Nullable
-	ListRow mLoadingLocation = null;
-
-	@Nullable
-	Service mLoadingBouquet;
-	@Nullable
-	Service mSelectedBouquet;
-	@Nullable
-	ServiceNowNext mSelectedService;
-
-	@Nullable
-	private Job mPrefetchJob;
-	private Job mBouquetJob;
-	@Nullable
-	private Job mServiceJob;
-	@Nullable
-	private Job mMovieJob;
-
-	private class BouquetHeaderItem extends HeaderItem {
-		private Service mBouquet;
-
-		public BouquetHeaderItem(long id, @NonNull Service bouquet) {
-			super(id, bouquet.getName() != null && !bouquet.getName().isEmpty()
-					? bouquet.getName() : getString(R.string.not_available));
-			mBouquet = bouquet;
-		}
-
-		public Service bouquet() {
-			return mBouquet;
-		}
-	}
-
-	private class SettingsRow extends ListRow {
-		public SettingsRow(HeaderItem header, ObjectAdapter adapter) {
-			super(header, adapter);
-		}
-	}
-
-	private class ServiceRow extends ListRow {
-		public ServiceRow(HeaderItem header, ObjectAdapter adapter) {
-			super(header, adapter);
-		}
-	}
-
-	private class MovieRow extends ListRow {
-		public MovieRow(HeaderItem header, ObjectAdapter adapter) {
-			super(header, adapter);
-		}
-	}
-
-	public RootBrowseFragment() {
-	}
-
-	@Override
-	public void onCreate(Bundle savedInstanceState) {
-		setRetainInstance(true);
-		super.onCreate(savedInstanceState);
-		setOnItemViewClickedListener(this);
-		setOnItemViewSelectedListener(this);
-		if (mBouquets == null)
-			mBouquets = new ArrayList<>();
-		if (mBouquetQueue == null)
-			mBouquetQueue = new ArrayList<>();
-		if (mLocations == null)
-		 	mLocations = new HashMap<>();
-
-		setHeadersState(HEADERS_ENABLED);
-		setBrandColor(ContextCompat.getColor(getContext(), R.color.primary_dreamdroid));
-		setBadgeDrawable(ContextCompat.getDrawable(getContext(), R.drawable.dreamdroid_header));
-		addSettingsRow();
-		mRequireReload = true;
-	}
-
-	@Override
-	public void onPause() {
-		super.onPause();
-		cancelBouquetAndServiceLoads();
-	}
-
-	@Override
-	public void onResume() {
-		super.onResume();
-		mRequireReload = mRequireReload || mBouquets == null || mBouquets.isEmpty() || mLocations == null || mLocations.isEmpty();
-		if (mRequireReload)
-			load();
-	}
-
-	@Override
-	public void onStart() {
-		super.onStart();
-		DreamDroid.setCurrentProfileChangedListener(this);
-	}
-
-	@Override
-	public void onStop() {
-		DreamDroid.setCurrentProfileChangedListener(null);
-		super.onStop();
-	}
-
-	private void cancelBouquetAndServiceLoads() {
-		if (mPrefetchJob != null) {
-			mPrefetchJob.cancel(null);
-			mPrefetchJob = null;
-		}
-		if (mServiceJob != null) {
-			mServiceJob.cancel(null);
-			mServiceJob = null;
-		}
-		if (mBouquetJob != null) {
-			mBouquetJob.cancel(null);
-			mBouquetJob = null;
-		}
-	}
-
-	private void cancelMovieLoad() {
-		if (mMovieJob != null) {
-			mMovieJob.cancel(null);
-			mMovieJob = null;
-		}
-	}
-
-	private void cancelAllLoads() {
-		cancelBouquetAndServiceLoads();
-		cancelMovieLoad();
-	}
-
-	protected void load() {
-		mRequireReload = false;
-		cancelAllLoads();
-		resetRows();
-		Toast.makeText(getContext(), R.string.loading, Toast.LENGTH_LONG).show();
-		mBouquetQueue.clear();
-		mBouquets.clear();
-		mLocations.clear();
-		mListRowAdapters.clear();
-		mLoadingBouquet = null;
-		mLoadingLocation = null;
-		ArrayList<NameValuePair> params = new ArrayList<>();
-		params.add(new NameValuePair("bRef", BOUQUETS_TV));
-		if (!isAdded() || getView() == null) {
-			mRequireReload = true;
-			return;
-		}
-		final ArrayList<NameValuePair> bouquetParams = params;
-		mPrefetchJob = TvBrowsePrefetchKt.launchTvBrowsePrefetch(this, () -> {
-			if (!isAdded() || getView() == null) {
-				return kotlin.Unit.INSTANCE;
-			}
-			mBouquetJob = ServiceListLoadKt.launchServiceListLoad(this, bouquetParams,
-					(success, services, errorText) -> {
-						onBouquetsReady(success, services, errorText);
-						return kotlin.Unit.INSTANCE;
-					});
-			return kotlin.Unit.INSTANCE;
-		});
-	}
-
-	protected void reload() {
-
-	}
-
-	private void onBouquetsReady(boolean success,
-			@NonNull List<net.reichholf.dreamdroid.enigma.Service> services,
-			@Nullable String errorText) {
-		if (!isAdded()) {
-			return;
-		}
-		if (!success) {
-			if (errorText != null) {
-				Toast.makeText(getContext(), errorText, Toast.LENGTH_LONG).show();
-			}
-			return;
-		}
-		onLoadBouquetsFinished(new ArrayList<>(services));
-	}
-
-	protected void resetRows() {
-		//Remove everything but settings
-		mRowsAdapter.removeItems(0, mRowsAdapter.size() - 1);
-	}
-
-	protected void addSettingsRow() {
-		ArrayObjectAdapter listRowAdapter = new ArrayObjectAdapter(new CardPresenter(CardPresenter.ItemMode.MODE_IMAGE));
-
-		listRowAdapter.add(new BrowseItem.Settings(BrowseItem.Kind.Reload,
-				getString(R.string.reload), R.drawable.ic_badge_reload));
-		listRowAdapter.add(new BrowseItem.Settings(BrowseItem.Kind.Preferences,
-				getString(R.string.settings), R.drawable.ic_badge_settings));
-		listRowAdapter.add(new BrowseItem.Settings(BrowseItem.Kind.Profile,
-				getString(R.string.profile), R.drawable.ic_badge_profiles));
-
-		HeaderItem header = new HeaderItem(ROW_SETTINGS_ID, getString(R.string.preferences));
-		mRowsAdapter.add(new SettingsRow(header, listRowAdapter));
-	}
-
-	public void onLoadBouquetsFinished(@NonNull ArrayList<Service> bouquets) {
-		mBouquets.addAll(bouquets);
-		mBouquetQueue.addAll(bouquets);
-
-		for (String location : DreamDroid.getLocations()) {
-			mLocations.put(location, new ArrayList<>());
-		}
-
-		Collections.reverse(mBouquetQueue);
-		loadNextBouquet();
-	}
-
-	public void loadNextBouquet() {
-		if (mBouquetQueue.isEmpty()) {
-			setSelectedPosition(0);
-			addLocations();
-			return;
-		}
-		mLoadingBouquet = mBouquetQueue.get(0);
-		mBouquetQueue.remove(0);
-		ArrayList<NameValuePair> params = new ArrayList<>();
-		params.add(new NameValuePair("bRef", mLoadingBouquet.getReference()));
-		if (!isAdded() || getView() == null) {
-			return;
-		}
-		if (mServiceJob != null) {
-			mServiceJob.cancel(null);
-			mServiceJob = null;
-		}
-		mServiceJob = EpgNowNextLoadKt.launchEpgNowNextLoad(this, params, (success, rows, errorText) -> {
-			onServicesReady(success, rows, errorText);
-			return kotlin.Unit.INSTANCE;
-		});
-	}
-
-	private void onServicesReady(boolean success, @NonNull List<ServiceNowNext> rows,
-			@Nullable String errorText) {
-		if (!isAdded()) {
-			return;
-		}
-		if (!success) {
-			if (errorText != null) {
-				Toast.makeText(getContext(), errorText, Toast.LENGTH_LONG).show();
-			}
-			loadNextBouquet();
-			return;
-		}
-		onLoadServicesFinished(new ArrayList<>(rows));
-	}
-
-	protected void addLocations() {
-		for(String location : mLocations.keySet()) {
-			ArrayObjectAdapter listRowAdapter = new ArrayObjectAdapter(new CardPresenter(CardPresenter.ItemMode.MODE_TEXT));
-			HeaderItem header = new HeaderItem(ROW_MOVIE_OFFSET_ID + mRowsAdapter.size(), location);
-			//Insert before prefs
-			mRowsAdapter.add(mRowsAdapter.size() - 1, new MovieRow(header, listRowAdapter));
-			mListRowAdapters.add(listRowAdapter);
-		}
-	}
-
-	public void onLoadServicesFinished(@NonNull ArrayList<ServiceNowNext> services) {
-		ArrayObjectAdapter listRowAdapter = new ArrayObjectAdapter(new CardPresenter(CardPresenter.ItemMode.MODE_IMAGE));
-		for (ServiceNowNext service : services) {
-			listRowAdapter.add(new BrowseItem.Service(service));
-		}
-		HeaderItem header = new BouquetHeaderItem(mRowsAdapter.size(), mLoadingBouquet);
-		mRowsAdapter.add(0, new ServiceRow(header, listRowAdapter));
-		mListRowAdapters.add(listRowAdapter);
-		loadNextBouquet();
-	}
-
-	protected void onLoadMoviesFinished(@NonNull ArrayList<net.reichholf.dreamdroid.enigma.Movie> movies) {
-		if (mLoadingLocation == null)
-			return;
-		ListRow listRow = (ListRow) mRowsAdapter.get( mRowsAdapter.indexOf(mLoadingLocation) );
-		ArrayObjectAdapter listRowAdapter = (ArrayObjectAdapter) listRow.getAdapter();
-		listRowAdapter.clear();
-		ArrayList<net.reichholf.dreamdroid.enigma.Movie> locs = mLocations.get(mLoadingLocation.getHeaderItem().getName());
-		for (net.reichholf.dreamdroid.enigma.Movie movie : movies) {
-			listRowAdapter.add(new BrowseItem.Movie(movie));
-			locs.add(movie);
-		}
-	}
-
-	private void onMoviesReady(boolean success, @NonNull List<net.reichholf.dreamdroid.enigma.Movie> movies,
-			@Nullable String errorText) {
-		if (!isAdded()) {
-			return;
-		}
-		if (!success) {
-			if (errorText != null) {
-				Toast.makeText(getContext(), errorText, Toast.LENGTH_LONG).show();
-			}
-			return;
-		}
-		onLoadMoviesFinished(new ArrayList<>(movies));
-	}
-
-
-
-	@Override
-	public void onItemClicked(Presenter.ViewHolder itemViewHolder, Object it, RowPresenter.ViewHolder rowViewHolder, Row row) {
-		if (it instanceof BrowseItem.Settings) {
-			BrowseItem.Settings settings = (BrowseItem.Settings) it;
-			if (settings.getKind() == BrowseItem.Kind.Reload) {
-				load();
-				return;
-			}
-			String type = settings.getKind() == BrowseItem.Kind.Preferences
-					? PreferenceActivity.PREFS_TYPE_GENERIC : PreferenceActivity.PREFS_TYPE_PROFILE;
-			Intent intent = new Intent(getContext(), PreferenceActivity.class);
-			intent.putExtra(PreferenceActivity.KEY_PREFS_TYPE, type);
-			startActivityForResult(intent, REQUEST_CODE_PROFILE);
-			return;
-		}
-		if (it instanceof BrowseItem.Movie) {
-			net.reichholf.dreamdroid.enigma.Movie movie = ((BrowseItem.Movie) it).getMovie();
-			ExtendedHashMap map = MovieListMapperKt.movieToExtendedHashMap(movie);
-			startActivity(IntentFactory.getStreamFileIntent(
-					getActivity(), movie.getReference(), movie.getFileName(), movie.getTitle(), map));
-			return;
-		}
-		if (it instanceof BrowseItem.Service) {
-			ServiceNowNext service = ((BrowseItem.Service) it).getRow();
-			ExtendedHashMap map = ServiceListMapperKt.serviceNowNextToExtendedHashMap(service);
-			String bouquetRef = mSelectedBouquet != null ? mSelectedBouquet.getReference() : null;
-			String title = service.getNow() != null && !service.getNow().getTitle().isEmpty()
-					? service.getNow().getTitle() : service.getServiceName();
-			startActivity(IntentFactory.getStreamServiceIntent(
-					getActivity(), service.getServiceReference(), title, bouquetRef, map));
-		}
-	}
-
-
-	@Override
-	public void onItemSelected(Presenter.ViewHolder itemViewHolder, Object it, RowPresenter.ViewHolder rowViewHolder, Row row) {
-		mLoadingLocation = null;
-		mSelectedService = null;
-		mSelectedBouquet = null;
-
-		if (row instanceof MovieRow) {
-			String dirname = row.getHeaderItem().getName();
-			if (mLocations.get(dirname).isEmpty()) {
-				ArrayList<NameValuePair> params = new ArrayList<>();
-				params.add(new NameValuePair("dirname", dirname));
-				mLoadingLocation = (ListRow) row;
-				if (mLoadingLocation != null && isAdded() && getView() != null) {
-					cancelMovieLoad();
-					mMovieJob = MovieListLoadKt.launchMovieListLoad(this, params, (success, movies, errorText) -> {
-						onMoviesReady(success, movies, errorText);
-						return kotlin.Unit.INSTANCE;
-					});
-				}
-			}
-			return;
-		} else if (row instanceof ServiceRow) {
-			if (it instanceof BrowseItem.Service) {
-				mSelectedService = ((BrowseItem.Service) it).getRow();
-			}
-			BouquetHeaderItem headerItem = (BouquetHeaderItem) row.getHeaderItem();
-			mSelectedBouquet = headerItem.bouquet();
-		}
-	}
-
-	@Override
-	public void onProfileChanged(Profile p) {
-		if(isResumed())
-			load();
-		else
-			mRequireReload = true;
-	}
-}
-
